+++ default.nix
@@ -1,5 +1,6 @@
 { mkDerivation
 , aeson
+, base64-bytestring
 , bifunctors
 , data-default
 , ghc
@@ -19,15 +20,18 @@
 , these
 , time
 , transformers
+, uri-bytestring
 , webkitgtk3-javascriptcore
+, web-routes
 }:
 
 mkDerivation {
   pname = "reflex-dom-contrib";
-  version = "0.4";
+  version = "0.5.1";
   src = builtins.filterSource (path: type: baseNameOf path != ".git") ./.;
   buildDepends = [
     aeson
+    base64-bytestring
     bifunctors
     data-default
     ghcjs-base
@@ -46,6 +50,8 @@ mkDerivation {
     these
     time
     transformers
+    uri-bytestring
+    web-routes
   ];
   license = null;
 }