+++ src/Reflex/Dom/Contrib/Router.hs
@@ -3,6 +3,7 @@
 {-# LANGUAGE FlexibleContexts         #-}
 {-# LANGUAGE ForeignFunctionInterface #-}
 {-# LANGUAGE JavaScriptFFI            #-}
+{-# LANGUAGE LambdaCase               #-}
 {-# LANGUAGE OverloadedStrings        #-}
 {-# LANGUAGE RankNTypes               #-}
 {-# LANGUAGE RecursiveDo              #-}
@@ -10,74 +11,183 @@
 {-# LANGUAGE TemplateHaskell          #-}
 {-# LANGUAGE TypeFamilies             #-}
 
-module Reflex.Dom.Contrib.Router where
+module Reflex.Dom.Contrib.Router (
+  -- == High-level routers
+    webRoute
+  , partialPathRoute
+  , fullUriRoute
+  , route
+
+  -- == Router configuration
+  , RouteConfig(..)
+
+  -- == Routing result
+  , Route(..)
+
+  , uriOrigin
+
+  -- == Lenses
+  , routeConfig_forward
+  , routeConfig_back
+  , routeConfig_pushState
+  , route_value
+  , route_fullURI
+
+  -- = Low-level URL bar access
+  , getLocation'
+  , getUrlText'
+  ) where
 
 ------------------------------------------------------------------------------
-import           Control.Monad.IO.Class    (MonadIO)
+import           Control.Lens              (makeLenses)
+import           Control.Monad.Except      (ExceptT(..), lift, runExceptT)
+import           Control.Monad.IO.Class    (MonadIO, liftIO)
+import           Data.Bifunctor            (first)
+import           Data.Maybe                (fromMaybe)
+import           Data.Monoid               ((<>))
 import           Data.Default
-import           Data.Text                 (Text)
-import           Reflex.Dom                hiding (Window)
+import qualified Data.Text                 as T
+import qualified Data.Text.Encoding        as T
+import           GHCJS.DOM.Types           (Location(..) )
+import           Reflex.Dom                hiding (EventName, Window)
+import qualified URI.ByteString            as U
+import qualified Web.Routes.PathInfo       as WR
 #if ghcjs_HOST_OS
-import           Data.Maybe                (fromJust)
-import           Control.Monad.IO.Class    (liftIO)
 import qualified GHCJS.DOM                 as DOM
 import qualified GHCJS.DOM.Document        as DOM
-import qualified GHCJS.DOM.EventM          as DOM
+import           GHCJS.DOM.EventM          (on)
 import           GHCJS.DOM.History         (back, forward, pushState)
-import           GHCJS.DOM.Location        (toString)
+import           GHCJS.DOM.Location        (getPathname, toString)
 import           GHCJS.DOM.Window          (Window, getHistory,
                                             getLocation, popState)
 import           GHCJS.Marshal.Pure
 #else
+import           Control.Monad.Reader      (ReaderT)
 #endif
 
 ------------------------------------------------------------------------------
-data RouteConfig t = RouteConfig
+data RouteConfig t a = RouteConfig
   { _routeConfig_forward   :: Event t () -- ^ Move the browser history forward
   , _routeConfig_back      :: Event t () -- ^ Move the browser history back
-  , _routeConfig_pushState :: Event t Text -- ^ Push to the URL state
-  -- , _routeConfig_pathBase  :: Text
-  --   -- ^ The part of the URL not related to SPA routing
+  , _routeConfig_pushState :: Event t a  -- ^ Push to the URL state
   }
 
-instance Reflex t => Default (RouteConfig t) where
+makeLenses ''RouteConfig
+
+instance Reflex t => Default (RouteConfig t a) where
   def = RouteConfig never never never
 
-instance Reflex t => Monoid (RouteConfig t) where
+
+instance Reflex t => Monoid (RouteConfig t a) where
   mempty = def
   mappend (RouteConfig f1 b1 p1) (RouteConfig f2 b2 p2) =
     RouteConfig (mappend f1 f2) (mappend b1 b2) (leftmost [p1, p2])
 
-  mconcat rcs =
-    RouteConfig (mconcat $ map _routeConfig_forward rcs)
-                (mconcat $ map _routeConfig_back rcs)
-                (leftmost $ map _routeConfig_pushState rcs)
 
-data Route t = Route {
-    _route_value :: Dynamic t Text -- ^ URL value
+data Route t a = Route {
+    _route_value   :: Dynamic t (Either T.Text a)     -- ^ Routing value
+  , _route_fullURI :: Dynamic t (U.URIRef U.Absolute) -- ^ Full URI
   }
 
-instance HasValue (Route t) where
-  type Value (Route t) = Dynamic t Text
+makeLenses ''Route
+
+instance HasValue (Route t a) where
+  type Value (Route t a) = Dynamic t (Either T.Text a)
   value = _route_value
 
--- | Manipulate and track the URL text for dynamic routing of a widget
-route :: (HasWebView m, MonadWidget t m) => RouteConfig t -> m (Route t)
-#if ghcjs_HOST_OS
-route (RouteConfig goForward goBack sSet) = do
-  win <- askDomWindow
-  loc <- getLocation' win
+-------------------------------------------------------------------------------
+-- | Manipulate and track the URL 'GHCJS.DOM.Types.Location' for dynamic
+--   routing of a widget
+route
+  :: (HasWebView m, MonadWidget t m)
+  => (Location -> IO (Either T.Text a))
+     -- ^ Decode the part of the path beyond '_routeConfig_pathBase' into an a
+  -> (a -> IO T.Text)
+     -- ^ Encode the routing value to a Text that will be passed to `pushState`
+  -> RouteConfig t a
+    -- ^ Routing widget configuration
+  -> m (Route t a)
+route to from (RouteConfig goForward goBack sSet) = do
+  win     <- askDomWindow
+  loc0    <- liftIO $ windowURI win
+  locVal0 <- liftIO $ to =<< fmap (fromMaybe (error "No location"))
+                                  (getLocation win)
+
   Just hist <- liftIO $ getHistory win
   performEvent_ $ ffor goForward $ \_ -> liftIO (forward hist)
   performEvent_ $ ffor goBack    $ \_ -> liftIO (back hist)
-  setLoc <- performEvent $ ffor sSet $ \t -> do
-    pushState hist (pToJSVal (0 :: Int)) ("" :: Text) t
-    getLocation' win
-  newLocs <- getPopState
-  Route <$> holdDyn loc (leftmost [setLoc, newLocs])
-#else
-route = error "route is only available to ghcjs"
-#endif
+
+  _ <- performEvent $ ffor sSet $ \t -> liftIO $ do
+    s <- from t
+    pushState hist (pToJSVal (0 :: Int)) ("" :: T.Text) s
+    dispatchEvent'
+
+  locVals <- getPopState to
+  locs <- performEvent $ ffor locVals $ \_ -> liftIO (windowURI win)
+
+  Route <$> holdDyn locVal0 (leftmost [Right <$> sSet, locVals]) <*> holdDyn loc0 locs
+
+
+windowURI :: Window -> IO (U.URIRef U.Absolute)
+windowURI w = do
+  l <- fromMaybe (error "Window has no location") <$> getLocation w
+  either (error "No parse of window location") id . U.parseURI U.laxURIParserOptions . T.encodeUtf8 <$> toString l
+
+-------------------------------------------------------------------------------
+-- | Route a single page app according to any 'WR.PathInfo' a => a
+webRoute
+  :: (MonadWidget t m, WR.PathInfo a)
+  => T.Text     -- ^ The part of the URL not related to SPA routing
+  -> RouteConfig t a
+  -> m (Route t a)
+webRoute pathBase = route decoder (return . (pathBase <>) . WR.toPathInfo)
+  where
+    decoder l = runExceptT $ do
+      pn <- lift $ getPathname l
+      ExceptT . return $ first T.pack . WR.fromPathInfo . T.encodeUtf8 =<<
+        note (pfxErr pn pathBase) (T.stripPrefix pathBase pn)
+
+
+-------------------------------------------------------------------------------
+-- | Route a single page app according to the part of the path after
+--   pathBase
+partialPathRoute
+  :: MonadWidget t m
+  => T.Text    -- ^ The part of the URL not related to SPA routing
+  -> RouteConfig t T.Text
+  -> m (Route t T.Text)
+partialPathRoute pathBase = route decoder (return . (pathBase <>))
+  where
+    decoder l = (\pn -> maybe (Left $ pfxErr pn pathBase)
+                              (Right . id) (T.stripPrefix pathBase pn)
+                ) <$> getPathname l
+
+
+-------------------------------------------------------------------------------
+-- | Route a single page app according to the full URI
+--   The dynamic returned to you will be a Text of the full URI
+--   The route string you pass in will be forwarded as-is to
+--   pushState.
+fullUriRoute
+  :: MonadWidget t m
+  => RouteConfig t (U.URIRef U.Absolute)
+  -> m (Route t (U.URIRef U.Absolute))
+fullUriRoute cfg = do
+  route uriDecoder (return . T.decodeUtf8 . U.serializeURIRef') cfg
+    where
+      uriDecoder = fmap (first (T.pack . show) .
+                         U.parseURI U.laxURIParserOptions .
+                         T.encodeUtf8) .
+                   toString
+
+
+uriOrigin :: U.URIRef U.Absolute -> T.Text
+uriOrigin r = T.decodeUtf8 $ U.serializeURIRef' r'
+  where
+    r' = r { U.uriPath = mempty
+           , U.uriQuery = mempty
+           , U.uriFragment = mempty
+           }
 
 #if ghcjs_HOST_OS
 -- | Get the DOM window object.
@@ -92,57 +202,43 @@ askDomWindow :: (MonadIO m) => m Window
 askDomWindow = error "askDomWindow is only available to ghcjs"
 #endif
 
-getLocation' :: MonadIO m => Window -> m Text
+
+getPopState :: (MonadWidget t m) => (Location -> IO (Either T.Text a)) -> m (Event t (Either T.Text a))
+getPopState decodeLocation = do
+  window <- askDomWindow
+  wrapDomEventMaybe window (`on` popState) $ do
+    liftIO $ getLocation window >>= \case
+      Nothing -> return Nothing
+      Just l  -> Just <$> decodeLocation l
+
+
+-------------------------------------------------------------------------------
+-- | (Unsafely) get the 'GHCJS.DOM.Location.Location' of a window
+getLocation' :: (HasWebView m, MonadIO m) => m Location
 #if ghcjs_HOST_OS
-getLocation' w = toString . fromJust =<< liftIO (getLocation w)
+getLocation' = fromMaybe (error "Window has no Location") <$>
+  (askDomWindow >>= liftIO . getLocation)
 #else
 getLocation' = error "getLocation' is only available to ghcjs"
 #endif
 
-getPopState :: (MonadWidget t m) => m (Event t Text)
-#if ghcjs_HOST_OS
-getPopState = do
-  window <- askDomWindow
-  wrapDomEventMaybe window (`DOM.on` popState) $ do
-    l <- getLocation window
-    case l of
-      Nothing -> return Nothing
-      Just loc -> do t <- toString loc; return (Just t)
-#else
-getPopState = error "getPopState is only available to ghcjs"
-#endif
 
-setWindowUrl :: MonadWidget t m => Event t Text -> m ()
-#if ghcjs_HOST_OS
-setWindowUrl url = do
-  performEvent_ $ ffor url $ \u -> do
-    win <- askDomWindow
-    Just hist <- liftIO $ getHistory win
-    pushState hist (pToJSVal (0 :: Int)) ("" :: Text) u
-#else
-setWindowUrl = error "setWindowUrl only available to ghcjs"
-#endif
+-------------------------------------------------------------------------------
+-- | (Unsafely) get the URL text of a window
+getUrlText' :: (HasWebView m, MonadIO m) => m T.Text
+getUrlText' = getLocation' >>= liftIO . toString
 
-getWindowInitUrl :: MonadWidget t m => m Text
-getWindowInitUrl = getLocation' =<< askDomWindow
-
-getWindowUrl :: MonadWidget t m => m (Dynamic t Text)
-getWindowUrl = do
-  win <- askDomWindow
-  loc <- getLocation' win
-  newLocs <- getPopState
-  holdDyn loc newLocs
 
 #if ghcjs_HOST_OS
+foreign import javascript unsafe "w = window; e = new PopStateEvent('popstate',{'view':window,'bubbles':true,'cancelable':true}); w.dispatchEvent(e);"
+  dispatchEvent' :: IO ()
 #else
-data Document
-data Location
 data Window
 data JSVal
 data History
 
-data SerializedScriptValue =
-  SerializedScriptValue { unSerializedScriptValue :: JSVal }
+dispatchEvent' :: IO ()
+dispatchEvent' = undefined
 
 forward :: History -> IO ()
 forward = undefined
@@ -150,26 +246,42 @@ forward = undefined
 back :: History -> IO ()
 back = undefined
 
-class FromJSVal a where
-  fromJSVal :: JSVal -> IO (Maybe a)
-
 getLocation :: Window -> IO (Maybe Location)
 getLocation = undefined
+
+getPathname :: Location -> IO T.Text
+getPathname = undefined
 
 getHistory :: Window -> IO (Maybe History)
 getHistory = undefined
 
-getState :: History -> IO (Maybe SerializedScriptValue)
-getState = undefined
-
-toString :: Location -> IO Text
-toString = undefined
-
-getDefaultView :: Document -> IO (Maybe Window)
-getDefaultView = undefined
-
-pushState :: History -> JSVal -> Text -> Text -> IO ()
+pushState :: History -> JSVal -> T.Text -> T.Text -> IO ()
 pushState = undefined
 
+popState :: EventName Window PopStateEvent
+popState = undefined
 
+pToJSVal :: Int -> JSVal
+pToJSVal = undefined
+
+on :: Window -> EventName t e -> EventM t e () -> IO (IO ())
+on = undefined
+
+type EventM t e = ReaderT e IO
+data PopStateEvent
+data EventName t e
+
+toString :: Location -> IO T.Text
+toString = undefined
 #endif
+
+-------------------------------------------------------------------------------
+-- | Helper functions
+note :: e -> Maybe a -> Either e a
+note _ (Just a) = Right a
+note e _        = Left e
+
+
+pfxErr :: T.Text -> T.Text -> T.Text
+pfxErr pn pathBase = "Encountered path (" <> pn
+            <> ") without expected prefix (" <> pathBase <> ")"