+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.5.1
+
+* Add Monoid instance for RouteConfig
+
+# 0.5
+
+* Add Route, webRoute, textRoute
+
 # 0.4.2
 
 * Add DynamicList widget